feat: move leave flow to ··· button, fix UX polish
- Move "Leave household" into a bottom sheet triggered by ··· button in
  the household card header (HouseholdPanel) instead of a footer button
- Move HouseholdPromo inside MembersScreen so it renders above the
  100px bottom padding and is actually visible
- Remove leave flow state/logic from MembersScreen (now in HouseholdPanel)
- Add 10px top padding to GroceryPage quick-add row for breathing room
- Add RD-09 through RD-16 to BACKLOG.md (push notifs, grocery affiliate
  links, photo attachments, activity feed, streaks, reactions, guest
  preview, WhatsApp invite)
